Style inline code with foreground instead of background colour
Because pandoc uses "white-space: pre-wrap" for <code> as per the CommonMark
recommendation, there can be whitespace at the end of a line if a code
span runs across a linebreak. This is visible when using a background
colour for code, and ugly, hence switching to using a foreground colour
for inline code instead.