Fix bibliography heading (#204)
After upgrading to pandoc 3.5, the bibliography heading stopped being a
heading due to a change in
jgm/pandoc@3d902349d1f8fd9a54539f8e16772c781dbd1063. I thought the
behaviour was a bug and filed jgm/pandoc#10367, but it's behaviour as
expected, and the workaround is simple: just add a heading manually.
This PR removes the metadata setting for the reference section heading,
along with documentation and test updates.